\name{rosio}
\alias{rosio}


\alias{dataset}
\alias{device}
\alias{devices}
\alias{getJWTtoken}
\alias{jwt        }
\alias{jwt<-}
\alias{key}
\alias{key<-}
\alias{messages}
\alias{password}
\alias{password<-}
\alias{Rosio}
\alias{username}
\alias{username<-}
\alias{whoami}

\title{An R interface to OpenSensors.io}


\description{

  This is currently only a stub. Documentation will follow when the
  package has reached some stability. It's still very early days, I am
  still experimenting and the code is still subject to major changes.

  To report issues, request features or contribute, please use the
  package's github repository. The vignette provides an overview of the
  current functionality.
}

\url{https://github.com/lgatto/rosio}